add documentation for making transit light curves; hide `batman_transit` for clarify
