+++ chromatic/__init__.py
@@ -1,6 +1,6 @@
 from .version import *
 from .rainbows import *
 from .imports import *
-from .resampling import *
 from .spectra import *
 from .archives import *
+from .tools import *
