+++ chromatic/rainbows/readers/coulombe.py
@@ -1,6 +1,7 @@
 """
 Define a reader for STScI pipeline x1dints.fits files.
 """
+
 from ...imports import *
 
 __all__ = ["from_coulombe"]
