+++ chromatic/rainbows/writers/xarray_fitted_light_curves.py
@@ -1,6 +1,6 @@
 """
-    3. Edit the `to_xarray_fitted_light_curves` function so that it will
-    write out a Rainbow to whatever format you want.
+3. Edit the `to_xarray_fitted_light_curves` function so that it will
+write out a Rainbow to whatever format you want.
 """
 
 # import the general list of packages
