+++ chromatic/tools/resampling.py
@@ -3,7 +3,11 @@ Tools for resampling array from one grid
 of independent variables to another.
 """
 
-from .imports import *
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.interpolate import interp1d
+import astropy.units as u
+
 
 __all__ = [
     "bintoR",
